+++ px_device_identity/device/__init__.py
@@ -1,0 +1,7 @@
+'''Device'''
+
+from .classes import (
+    DeviceProperties,
+    DeviceRegistrationProperties
+)
+from .main import Device