+++ px_device_identity/device/host.py
@@ -1,0 +1,12 @@
+import platform
+import psutil
+
+def system_information():
+    uname = platform.uname()
+    memory = psutil.virtual_memory()[0]
+    return {
+        'operating_system': uname.system,
+        'operating_system_release': uname.release,
+        'system_architecture': uname.machine,
+        'system_memory': memory
+    }