+++ px_device_identity/device/jwk.py
@@ -3,22 +3,23 @@ from json import dumps as json_dumps
 from authlib.jose import jwk
 from exitstatus import ExitStatus
 
-from .log import Logger
-from .config import KEY_DIR
+from px_device_identity.log import Logger
 from .util import split_key_type
-from .classes import RequestedOperation
+from .config import KEY_DIR
 
 log = Logger(__name__)
 
 
 class JWK:
     '''Working with JWK(S)'''
-    def __init__(self, operation_class: RequestedOperation, key_dir=KEY_DIR):
-        self.security = vars(operation_class)['security']
-        self.key_type = vars(operation_class)['key_type']
+    def __init__(self, device_properties: 'DeviceProperties', key_dir=KEY_DIR):
+        self.key_security = device_properties.key_security
+        self.key_type = device_properties.key_type
         self.key_dir = key_dir
         self.jwk_path = key_dir + 'public_jwk.json'
         self.public_key_path = key_dir + 'public.pem'
+        '''The key is generated automatically'''
+        self.key = self._generate()
 
     def _generate(self):
         key_cryptography = split_key_type(self.key_type)[0]
@@ -44,24 +45,22 @@ class JWK:
 
     def save_to_key_path(self) -> True:
         '''Generates and saves JWK to the default path'''
-        key = self._generate()
-        formatted_key = bytearray(json_dumps(key, ensure_ascii=True).encode('utf8'))
+        formatted_key = bytearray(json_dumps(self.key, ensure_ascii=True).encode('utf8'))
         try:
             with open(self.jwk_path, 'wb') as writer:
                 writer.write(formatted_key)
-                return True
+            return True
         except:
             log.error('Could not save JWK to {}'.format(self.jwk_path))
             sys.exit(ExitStatus.failure)
 
     def get(self):
         '''Generates and returns JWK'''
-        return self._generate()
+        return self.key
 
     def get_jwks(self):
         '''Generates JWK and returns JWKS'''
-        key = self._generate()
         jwks = {
-            'keys': [key]
+            'keys': [self.key]
         }
         return jwks