+++ px_device_identity/main.py
@@ -1,33 +1,21 @@
+'''PantherX Device Identity'''
+
 import sys
-from getpass import getuser
-from json import dumps as json_dumps
 import pkg_resources
 
 from exitstatus import ExitStatus
 
-from .classes import DeviceClass
-from .device import Device
-from .jwk import JWK
 from .cli import get_cl_arguments
-from .sign import Sign
 from .log import Logger
-from .config import DeviceConfig
-from .migration import first_migration_key_dir, second_migration_add_config_key_domain
-
-import platform
-opsys = platform.system()
+# from .migration import first_migration_key_dir, second_migration_add_config_key_domain
+from .util import is_superuser_or_quit
+from .classes import OperationProperties
+from .device import Device, DeviceProperties
 
 log = Logger(__name__)
+
 version = pkg_resources.require("px_device_identity")[0].version
 
-
-def handle_result(success):
-    if success:
-        log.info("We're done here.")
-        sys.exit(ExitStatus.success)
-    else:
-        log.info("Something went wrong.")
-        sys.exit(ExitStatus.failure)
 
 def main():
     log.info('------')
@@ -35,69 +23,47 @@ def main():
     log.info('v{}'.format(version))
     log.info('------')
 
-    if opsys == 'Linux':
-        current_user = getuser()
-        if current_user != 'root':
-            log.warning('!!! This application is designed to run as root on the target device !!!')
-            log.warning('!!! Current user: {}'.format(current_user))
-            sys.exit()
-    if opsys == 'Windows':
-        import ctypes
-        is_windows_admin = False
-        try:
-            is_windows_admin = ctypes.windll.shell32.IsUserAnAdmin() == 1
-        except:
-            log.warning('!!! Something went wrong checking if the current user is an administrator')
-        if not is_windows_admin:
-            log.warning('!!! This application is designed to run as administrator on the target device !!!')
-            # sys.exit()
-
+    is_superuser_or_quit()
 
     cl_arguments = get_cl_arguments()
-    operation_class = cl_arguments.get('operation')
-    device_type: str = cl_arguments.get('device_type')
-    device_is_managed: bool = cl_arguments.get('device_is_managed')
-    message: str = cl_arguments.get('message')
-    host: str = cl_arguments.get('host')
-    domain: str = cl_arguments.get('domain')
-    location: str = cl_arguments.get('location')
+    operation: OperationProperties = cl_arguments['operation']
+    device_properties: DeviceProperties = cl_arguments['device_properties']
+    message: str = cl_arguments['message']
+    # debug: bool = cl_arguments['debug']
 
-    device_dict = DeviceClass(device_type, device_is_managed)
+    device = Device(operation.force_operation)
 
-    device = Device(operation_class, device_dict)
-    INITIATED = device.check_init()
-    if INITIATED is True:
-        '''Run required migrations'''
-        first_migration_key_dir()
-        second_migration_add_config_key_domain()
+    is_initiated = device.is_initiated
+    # if is_initiated is True:
+        # print('Supposed to run migration.')
+        # '''Run required migrations'''
+        # first_migration_key_dir()
+        # second_migration_add_config_key_domain()
 
-    if operation_class.action != 'INIT' and INITIATED is False:
+    if operation.action != 'INIT' and is_initiated is False:
         log.error('Device is not initiated.')
         log.error('Initiate device with --operation INIT --type <DEFAULT|TPM>')
         sys.exit(ExitStatus.failure)
 
-    if operation_class.action == 'INIT':
-        initiated = device.init(host, domain, location)
-        handle_result(initiated)
+    if operation.action == 'INIT':
+        try:
+            device.init(device_properties)
+        except:
+            print(sys.exc_info()[0])
+            log.info("Something went wrong.")
+            sys.exit(ExitStatus.failure)
 
-    config = DeviceConfig().get()
-    operation_class.security = config.get('keySecurity')
-    operation_class.key_type = config.get('keyType')
+        sys.exit(ExitStatus.success)
 
-    log.info('DEVICE CONFIG')
-    log.info(config)
+    # Here we initiate the device properties from config, since we are not initiating
+    if operation.action == 'GET_JWK':
+        return device.get_jwk()
 
-    if operation_class.action == 'GET_JWK':
-        jwk = JWK(operation_class)
-        return json_dumps(jwk.get())
+    if operation.action == 'GET_JWKS':
+        return device.get_jwks()
 
-    if operation_class.action == 'GET_JWKS':
-        jwk = JWK(operation_class)
-        return json_dumps(jwk.get_jwks())
-
-    if operation_class.action == 'SIGN':
-        sign = Sign(operation_class, message)
-        return sign.sign()
+    if operation.action == 'SIGN':
+        return device.sign(message)
 
 if __name__ == '__main__':
     main()